requirements.txt
pyyaml==5.4
jsonschema==3.0.2
SPARQLWrapper
requests==2.21.0
pyrsistent==0.16.0
python-frontmatter==0.5.0
yamllint==1.25.0
ruamel.yaml==0.16.12
